+++ packages/cli/vite.config.ts
@@ -1,5 +1,11 @@
 import { createKernelLibConfig } from '../../vite.config.base';
 
-export default createKernelLibConfig('@geekist/wp-kernel-cli', {
-	index: 'src/index.ts',
-});
+export default createKernelLibConfig(
+	'@geekist/wp-kernel-cli',
+	{
+		index: 'src/index.ts',
+	},
+	{
+		external: ['chokidar', 'clipanion', 'cosmiconfig', 'typanion'],
+	}
+);